Tutorial and Basic Game Setup
This morning I added some basic stuff, like making part of the tutorial level, making a bunch of sprites, revising the Samurai animations, and more. I do have an error with the animation player sadly for the attack animation on the player scene, but I'll be able to fix that when I can later.